requirements.txt
google-genai>=0.3.0
python-dotenv>=1.0.0
rich>=13.0.0
openai>=1.0.0
textual>=0.40.0
